requirements.txt
matplotlib==3.10.3
numpy==2.3.1
Pillow==11.2.1
scikit-learn==1.7.0
seaborn==0.13.2
SQLAlchemy==2.0.41
streamlit==1.45.1
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
